scripts/kallsyms: set relative_base more effectively
Currently, record_relative_base() iterates over the entire table to
find the minimum address, but it is not efficient because we sort
the table anyway.

After sort_symbol(), the table is sorted by address. (kallsyms parses
the 'nm -n' output, so the data is already sorted by address, but this
commit does not rely on it.)

Move record_relative_base() after sort_symbols(), and take the first
non-absolute symbol value.

Signed-off-by: Masahiro Yamada <yamada.masahiro@socionext.com>
Signed-off-by: CloudedQuartz <ravenklawasd@gmail.com>
Signed-off-by: nao04 <korewanao@outlook.com>
Signed-off-by: dodyirawan85 <40514988+dodyirawan85@users.noreply.github.com>